Hugepages Shared Memory Support in LTTng
This patch is not expected to be integrated into LTTng. Its purpose is
to show the viability and advantages of hugepages.

To use hugepages for the shared memory, we need to create a memory
mapped files on a hugetlbfs directory /dev/hugepages instead of
/dev/shm. See [1] for the instruction to mount a hugetlbfs virtual file
system.

This patch simply changes the location of the shm file and replaces
shm_open/shm_unlink with open/unlink. This patch could not work
independently. It should be used together with another patch in
lttng-ust [2]. See that patch for more hugepages information.

This patch is based on commit 4e328cceb6ff5fd6240c866b52b29234946c5c66.

[1] https://wiki.debian.org/Hugepages
[2] [RFC PATCH lttng-ust] Hugepages Shared Memory Support in LTTng

Signed-off-by: Yiteng Guo <guoyiteng@gmail.com>